Check for obsolete closed tasks without downloading artifacts
Koji doesn't give us a straightfoward method to check for any task
output.
In order to know whether there is something to download, we need to
find the descdendents of the original task, then look at the srpm build output.
If there's some, we're good. If there's none, we need to rebuild the
srpm.
This adds two new calls to koji, hence makes the script run last longer.
It's not really needed in the day-to-day operations, but after a
few-weeks-long pause the issue with trying to download non-existing srpms starts to appear.
Hence it's disabled by default, but available with a function flag.
If we ever add CLI options to the script, it could be propagated up,
ATM it'd be probably too much hassle.