Remove koji.id if it's older than one week
If the task is closed and we haven't downloaded the srpm yet,
before we attempt to do that, we need to check whether there are
artifacts in Koji.
We assume that Koji deletes artifacts a week after a successful
scratchbuild. If we detect that our koji.id is older than a week,
we can just retrigger the scratchbuild and save ourselves a failed
download of the srpm.