Fix readiness reporting with extras subpackages
This fixes the stderr output: When an extra package was checked first,
the unconditional break prevented the ✔ print.

Moreover, this fixes a problem where an extra was not yet ready, but it was required.

Previously, the code would happily report weasyprint as ready when fonttools was available,
but fonttools[woff] was not (weasyprint BRs fonttools[woff]).